Make temporal types ICypherValue and rename files for simplicity
