+++ Neo4j.Driver/Neo4j.Driver.Tests/IO/StructHandlers/Temporal/DateTimeHandlerTests.cs
@@ -15,7 +15,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System;
 using System.Collections;
 using System.Collections.Generic;
 using FluentAssertions;
@@ -30,46 +29,46 @@ using Xunit;
 
 namespace Neo4j.Driver.Tests.IO.StructHandlers
 {
-    public class CypherTimeWithOffsetHandlerTests : StructHandlerTests
+    public class DateTimeHandlerTests : StructHandlerTests
     {
-        internal override IPackStreamStructHandler HandlerUnderTest => new CypherTimeWithOffsetHandler();
+        internal override IPackStreamStructHandler HandlerUnderTest => new DateTimeHandler();
 
         [Fact]
-        public void ShouldWriteTimeWithOffset()
+        public void ShouldWriteDateTime()
         {
-            var time = new CypherTimeWithOffset(12, 35, 59, 128000987, (int)TimeSpan.FromMinutes(150).TotalSeconds);
+            var dateTime = new CypherDateTime(1978, 12, 16, 12, 35, 59, 128000987);
             var writerMachine = CreateWriterMachine();
             var writer = writerMachine.Writer();
 
-            writer.Write(time);
+            writer.Write(dateTime);
 
             var readerMachine = CreateReaderMachine(writerMachine.GetOutput());
             var reader = readerMachine.Reader();
 
             reader.PeekNextType().Should().Be(PackStream.PackType.Struct);
             reader.ReadStructHeader().Should().Be(2);
-            reader.ReadStructSignature().Should().Be((byte) 'T');
-            reader.Read().Should().Be(time.NanosecondsOfDay);
-            reader.Read().Should().Be((long)time.OffsetSeconds);
+            reader.ReadStructSignature().Should().Be((byte) 'd');
+            reader.Read().Should().Be(dateTime.EpochSeconds);
+            reader.Read().Should().Be((long)dateTime.NanosOfSecond);
         }
         
         [Fact]
-        public void ShouldReadTimeWithOffset()
+        public void ShouldReadDateTime()
         {
             var writerMachine = CreateWriterMachine();
             var writer = writerMachine.Writer();
 
-            writer.WriteStructHeader(CypherTimeWithOffsetHandler.StructSize, CypherTimeWithOffsetHandler.StructType);
-            writer.Write(45359128000987);
-            writer.Write((int)TimeSpan.FromMinutes(150).TotalSeconds);
+            writer.WriteStructHeader(DateTimeHandler.StructSize, DateTimeHandler.StructType);
+            writer.Write(1520919278);
+            writer.Write(128000987);
 
             var readerMachine = CreateReaderMachine(writerMachine.GetOutput());
             var reader = readerMachine.Reader();
             var value = reader.Read();
 
             value.Should().NotBeNull();
-            value.Should().BeOfType<CypherTimeWithOffset>().Which.NanosecondsOfDay.Should().Be(45359128000987);
-            value.Should().BeOfType<CypherTimeWithOffset>().Which.OffsetSeconds.Should().Be((int)TimeSpan.FromMinutes(150).TotalSeconds);
+            value.Should().BeOfType<CypherDateTime>().Which.EpochSeconds.Should().Be(1520919278);
+            value.Should().BeOfType<CypherDateTime>().Which.NanosOfSecond.Should().Be(128000987);
         }
         
     }
